fix26

+++ src/routing/comment-routing.ts
@@ -19,36 +19,38 @@ commentRouter.get('/:id', async (req: Request, res: Response) => {
 
 
 commentRouter.delete('/:commentId', authBearerMiddleware, async (req: Request, res: Response) => {
-    try {
-        const findComId = await commentRepository.getComId(req.params.commentId)
-        if (!findComId) {
-            res.sendStatus(404)
-        } else {
-            const deleteComId = await commentRepository.deleteCom(req.params.commentId)
 
-            res.sendStatus(204)
-            return
-        }
-    } catch {
+    const findComId = await commentRepository.getComId(req.params.commentId)
+    if (!findComId) {
+        res.sendStatus(404)
+        return
+    }
+    if (req.user!.id !== findComId?.commentatorInfo?.userId) {
         res.sendStatus(403)
 
+    } else {
+        const deleteComId = await commentRepository.deleteCom(req.params.commentId)
+        res.sendStatus(204)
+        return
     }
+
 
 })
 
 commentRouter.put('/:commentId', authBearerMiddleware, commentMiddleware, errorsMiddleware, async (req: Request, res: Response) => {
-    try {
-        const findComId = await commentRepository.getComId(req.params.commentId)
 
-        if (!findComId) {
-            res.sendStatus(404)
-
-        } else {
-            const comPut = await commentRepository.updateCom(req.body.content, req.params.commentId)
-            res.sendStatus(204)
-            return
-        }
-    } catch {
-        res.sendStatus(403)
+    const findComId = await commentRepository.getComId(req.params.commentId)
+    if (!findComId) {
+        res.sendStatus(404)
+        return
     }
+    if (req.user!.id !== findComId?.commentatorInfo?.userId) {
+        res.sendStatus(403)
+
+    } else {
+        const comPut = await commentRepository.updateCom(req.body.content, req.params.commentId)
+        res.sendStatus(204)
+        return
+    }
+
 })
